requirements.txt
scapy>=2.5.0
pyyaml>=6.0.0
numpy>=1.26.0
pandas>=2.2.0
scikit-learn>=1.4.0
rich>=13.0.0
sqlalchemy>=2.0.0
pydantic>=2.6.0
joblib>=1.3.0